Reestructuration of error request handlers in course.
Thinking of ErrorRH inheriting from StaticRH (that now needs a resource_path string for construction),
in order to use its assemble response_header and send_header methods.
I might need to include a protected constructor in StaticRH that does not need that string.

Or maybe just drop it and inherit form ARequestHandler. Those 2 method could be part of the the base class as well...